renamed controllers dir to config
